# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: openGym API
  version: 1.3.8
  summary: Passkey (WebAuthn) auth + per-user state storage for openGym
  description: |
    The backend of [openGym](https://opengym.duarte-santos.ch) — a free, open-source
    gym & body-weight tracker. One Node file, no framework, JSON-file storage,
    HMAC-signed session cookies.

    ### Authentication
    Sign-in is **passkey only** (WebAuthn, via `@simplewebauthn/server`). A successful
    `register/verify` or `login/verify` sets the session cookie; every authenticated
    route then works with that cookie. The paired mobile app has no shared cookie jar,
    so it redeems a short pairing code (`/api/pair/redeem`) for the *same* signed token
    and sends it as `Authorization: Bearer <token>` instead.

    ### Sessions
    The cookie/bearer value is `<payload>.<hmac>` where the payload is
    `<uid>:<expiry-ms>:<session-version>`, signed with a per-instance secret
    (HMAC-SHA256). Sessions last `SESSION_DAYS` (default 90). `POST /api/logout/all`
    bumps the user's session version, which invalidates every token ever issued for
    the account.

    ### CSRF
    State-changing browser requests must come from the app's own origin. The server
    checks `Sec-Fetch-Site` (falling back to `Origin` vs the configured `ORIGIN`);
    a mismatch is refused with `403 {"error":"cross-origin request refused"}` on any
    non-GET route. Exempt: the WebAuthn and `pair/redeem` routes (each carries its own
...

...

...
  license:
...
servers:
...
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /api/food/barcode:
    post:
      tags: [nutrition]
      operationId: foodBarcode
      summary: Look a product up in Open Food Facts by its barcode
      description: >-
        Asked from this server on the member's behalf and answered from its cache when possible
        (a found product for 90 days, a miss for 7). Off unless an admin switched online lookup
        on; `FOOD_LOOKUP_DISABLED=1` keeps it off regardless. What leaves the server is the
        barcode — never who asked. POST rather than GET so that what somebody eats does not end
        up in a URL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: { type: string, description: '8–14 digits; spaces and dashes are ignored.' }
                lang: { type: string, description: 'Preferred language of the product name (`ru` or anything else for English).' }
      responses:
        '200':
          description: The product, per 100 g.
          content: { application/json: { schema: { type: object, properties: { food: { $ref: '#/components/schemas/FoodRow' }, cached: { type: boolean } } } } }
        '400': { description: 'Not a barcode (`badcode`).', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { description: 'Online lookup is switched off (`disabled`).', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: 'Not in the database (`notfound`).', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '422': { description: 'The database has the product but no nutrition values for it (`nonutrition`).', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '429': { description: 'The per-member or instance-wide lookup limit was reached (`throttled`).', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '502': { description: 'Open Food Facts did not answer (`unreachable`). Not cached.', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/food/search:
    post:
      tags: [nutrition]
      operationId: foodSearch
      summary: Search Open Food Facts by product name
      description: >-
        Up to 20 products that have nutrition values. Meant to be asked on a button, never per
        keystroke — the database limits searches per address. Results are cached for a day.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [q]
              properties:
                q: { type: string, minLength: 3, maxLength: 60 }
                lang: { type: string }
      responses:
        '200':
          description: Matching products, per 100 g.
          content: { application/json: { schema: { type: object, properties: { items: { type: array, items: { $ref: '#/components/schemas/FoodRow' } }, cached: { type: boolean } } } } }
        '400': { description: 'Fewer than three characters (`short`).', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { description: 'Online lookup is switched off (`disabled`).', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '429': { description: 'Lookup limit reached (`throttled`).', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '502': { description: 'Open Food Facts did not answer (`unreachable`).', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: gymsid
      description: |
        HMAC-signed session cookie, set by `register/verify` / `login/verify`.
        On HTTPS deployments it is issued as **`__Host-gymsid`** (host-locked,
        Secure); the unprefixed `gymsid` is what a plain-http `localhost` instance
        uses, and is still accepted for pre-upgrade sessions. HttpOnly + SameSite=Lax;
        browser clients never handle the value directly.
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        The same signed session token, carried as `Authorization: Bearer <token>` by
        the paired mobile app (obtained from `POST /api/pair/redeem`). Bearer requests
        skip the CSRF origin check — a browser never attaches this header on its own.
    healthToken:
      type: http
      scheme: bearer
      description: >-
        A Health token (`ogh_…`), made by `POST /api/healthkit/token` and carried by an iOS
        Shortcut. Accepted by `POST /api/healthkit/ingest` only; it is not a session and reads
        nothing.
  responses:
    Unauthorized:
      description: No session, expired/revoked session, or a disabled account.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: "not signed in" }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable message. This is the only error shape the API has.
...
    FoodRow:
      type: object
      description: A product per 100 g, as the food diary stores one.
      required: [n, k, p, f, c]
      properties:
        code: { type: string, description: 'The barcode.' }
        n: { type: string, maxLength: 80 }
        brand: { type: string }
        k: { type: integer, description: 'Energy, kcal per 100 g.' }
        p: { type: number, description: 'Protein, g per 100 g.' }
        f: { type: number, description: 'Fat, g per 100 g.' }
        c: { type: number, description: 'Carbohydrate, g per 100 g.' }
        sv: { type: integer, description: 'A serving in grams or millilitres, when the database states one.' }
...